Fix test for followers
