+++ .github/workflows/build.yml
@@ -16,12 +16,11 @@ jobs:
         uses: actions/setup-python@v4
         with:
           python-version: ${{ matrix.python-version }}
-	  cache: 'pip'
+          cache: 'pip'
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
           make install-dev
-          if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
+      - uses: actions/checkout@v3
       - name: Run tests
-        run: |
-          make test
+        run: make test
